move flexible_writer/_reader into flexible.writer/.reader
